requirements.txt
logzero==1.5.0
neo-boa==0.3.7
neo-python-rpc==0.2.0
neocore==0.3.6